+++ packages/iocuak/container/services/api/ContainerBindingApiService.ts
@@ -5,5 +5,6 @@ export interface ContainerBindingApiService {
   bind<TInstance, TArgs extends unknown[]>(
     type: Newable<TInstance, TArgs>,
   ): void;
+  bindToValue<TInstance>(serviceId: ServiceId, value: TInstance): void;
   unbind(serviceId: ServiceId): void;
 }